refactor: add intermediate layer of learning units section on curriculum page
